+++ api/itineraries.json
@@ -17,6 +17,12 @@
       "eatery": "7",
       "attraction": "7",
       "id": 3
+    },
+    {
+      "park": "813548D1-F8EA-4C07-8878-17FB3C385378",
+      "eatery": "3",
+      "attraction": "4",
+      "id": 4
     }
   ]
 }
