+++ scripts/eateries/EateryDialog.js
@@ -3,14 +3,14 @@ import { useEateries } from './EateryDataProvider.js'
 const contentTarget = document.querySelector('.dialog__eatery')
 const eventHub = document.querySelector('.container')
 
-eventHub.addEventListener('eateryDetailBtnClicked', customEvent => {
+eventHub.addEventListener('eateryDetailBtnClicked', (customEvent) => {
   // Get the eatery id
   const eateryId = customEvent.detail.clickedEatery
 
   const eateryArray = useEateries()
 
   // Find returns the very first object that matches the condition in the callback function
-  const iFoundYou = eateryArray.find(currentEatery => {
+  const iFoundYou = eateryArray.find((currentEatery) => {
     if (currentEatery.id === parseInt(eateryId)) {
       return true
     }
@@ -21,7 +21,7 @@ eventHub.addEventListener('eateryDetailBtnClicked', customEvent => {
   contentTarget.showModal()
 })
 
-export const EateryDialog = eateryObject => {
+export const EateryDialog = (eateryObject) => {
   contentTarget.innerHTML = `
     <section class="eatery">
       <h4>${eateryObject.businessName}</h4>
