+++ scripts/weather/Weather.js
@@ -2,9 +2,7 @@ export const Weather = (weatherObject) => {
   return `
     <div class"weather">
       <h4>Weather</h4>
-      <p>weatherObject.Xthing</p>
+      <p>${weatherObject.description}</p>
     </div>
   `
 }
-
-// ${weatherObject.name}
